scon: tproxy: change nonstandard __u8 loop counter to int

+++ scon/bpf/src/tproxy.c
@@ -45,7 +45,7 @@ static __always_inline __u8 matches_subnet4(volatile __u32 ip, volatile __u32 su
 
 static __always_inline __u8 matches_subnet6(volatile __u32 *ip, volatile __u32 *subnet, volatile __u32 *mask) {
     #pragma clang loop unroll(enable)
-    for (__u8 i = 0; i < 4; i++) {
+    for (int i = 0; i < 4; i++) {
         if ((ip[i] & mask[i]) != subnet[i]) {
             return 0;
         }
